CAMBIOS REALIZADOS EN CLASE JUNTO A SEBASTIAN Se corrigio la forma de pasar la seleccion multiple del menu. Antes pasaba por Get, ahora pasa por Post por medio del ViewModel

+++ src/main/java/ar/edu/unlam/tallerweb1/dao/SalonDao.java
@@ -15,7 +15,7 @@ public interface SalonDao {
     List<Salon> buscarSalonesZonaNorte(Integer cantidad ,String fecha);
     List<Salon> buscarSalonesCapital(Integer cantidad ,String fecha);
 
-    Salon traerSalonPorId(Integer id);
+    Salon traerSalonPorId(Long id);
 
     void guardarReserva(Reserva reserva);
    Salon galeria(Integer id);
